requirements.txt
# I2I Refusal Bias Study - Dependencies
# IJCAI 2026

# Core
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
tqdm>=4.65.0

# Dataset
datasets>=2.14.0
huggingface-hub>=0.16.0

# Deep Learning
torch>=2.0.0
accelerate>=0.24.0
transformers>=4.55.0  # Required for Step1X-Edit-v1p2

# Diffusers - Install from GitHub for latest models
# pip install git+https://github.com/huggingface/diffusers
# For Step1X-Edit-v1p2:
# git clone -b step1xedit_v1p2 https://github.com/Peyton-Chen/diffusers.git
# cd diffusers && pip install -e .

# Evaluation
scipy>=1.11.0
statsmodels>=0.14.0

# VLM Evaluation
google-generativeai>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: RegionE for faster Step1X inference
# pip install RegionE

# Optional: bitsandbytes for FLUX.2 4-bit quantization
# bitsandbytes>=0.41.0
